feat: update run0edit to v0.4.3 w/ support for immutable flag (#1055)
* feat: update run0edit to v0.4.0 w/ support for immutable flag

The new version of `run0edit` checks for the immutable flag on the file
to be edited, and if so, asks the user whether to temporarily remove it
to edit the file. This makes editing immutable files less error-prone,
as the immutable flag is only removed for the brief window necessary to
overwrite the file contents and reapplied immediately afterward; the
script also compares the file contents after the immutable flag has been
reapplied to ensure it was not modified by another program during that
window.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>

* fix: handle case where lsattr gives error

* feat(run0edit): update to v0.4.1

* Handle immutable flag on directory.
* Check for read-only filesystems to error out earlier.

* improve control flow & message wording

* refactor how inner script is substituted into main script

No change to script functionality with this commit, but single-quotes no
longer need to be avoided in the inner script since it's wrapped in a
heredoc.

* split out duplicated code into functions

* simplify read-only filesystem check, reduce code duplication

---------

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>