fix(audit): only check permissions of flatpak apps, not runtimes (#1071)
* fix(audit): only check permissions of flatpak apps, not runtimes

Flatpak overrides on runtimes don't appear to impact the actual
permissions of flatpak apps using that runtime, so there's no point in
checking them. The behavior of the flatpak sandbox is determined by the
permissions on the app itself, which are already checked.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>

* fix: default to width 80 if can't get terminal width

For example this happens if the output is piped to another program.

---------

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>