ci: add workflow to run Ruff on Python code (#1095)
* ci: add workflow to run Ruff on Python code

Ruff is a Python linter/formatter: https://docs.astral.sh/ruff/
This workflow runs the Ruff linter on all Python code in the repo.

The added `ruff.toml` file configures Ruff's formatting settings; the
only change from the default is to set the maximum line length to 100,
which was already being used for the audit script. This helps ensure a
consistent formatting standard for Python code.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>

* ci: add format check, use ubuntu-24.04, run only on .py file changes

* ci(ruff): only run on pull_request, not push

---------

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>