feat(audit): merge kargs checks into single report (#1103)
* feat(audit): merge kargs checks into single report

This makes it so kernel arguments that are already set are not reported,
making the audit script's output significantly shorter for users who
have already run `ujust set-kargs-hardening`. Also adds a recommendation
to run `ujust set-kargs-hardening` if any expected kargs are missing.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>

* perf(audit): store current kargs as set instead of list

This should scale better if the user has a large number of kernel
arguments, since checking membership is O(1) instead of O(n).

* refactor: simplify control flow, don't unnecessarily build lists

---------

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>