fix: add retry logic for secureblue-flatpak-setup.service (#1257)
The retry logic is copied from securebluefirstrun.service and restarts
the service on failure, with exponential backoff in the event of
repeated failures.

Also add integration test for the service, and skip flatpak checks in
the audit service since that could check the flatpak setup service in
the middle of running and give timing-dependent results.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>