fix(test): add 5 second sleep to integration test (#1281)
The flatpak setup service is a user service, so it can't start until the
user logs in, which in the integration testing VM only happens when the
test user logs in via ssh to run the tests. This leaves only a very
brief time window for the service to start, so to give it more time, we
sleep for a few seconds at the start of the test.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>
Co-authored-by: RoyalOughtness <129108030+RoyalOughtness@users.noreply.github.com>