fix(securedns): restrict dns-selector menu for unknown DNS (#1424)
Currently, the shortened interactive dns-selector menu (i.e. without DNSSEC etc.)
is shown when systemd-resolved is in use.  Since detection for VPN DNS was added
(DNSResolver.UNKNOWN), this caused a bug where the extended list is mistakenly
shown for VPN DNS as well. This commit shows the short menu whenever Unbound is
not in use.