fix: move sysctl conf file to /usr/lib/sysctl.d (#1446)
`/usr/lib/sysctl.d` is the recommended location for sysctl configuration
set by the distribution. This also ensures that local changes in
`/etc/sysctl.d` won't prevent the user from receiving updates to the
secureblue sysctl configuration.

Renaming the file from `60-hardening.conf` to `55-hardening.conf` is to
ensure that this file will be lower priority but not overridden by
`/etc/sysctl.d/60-hardening.conf` for users who already changed that
file (and therefore will not have `60-hardening.conf` erased with this
update).

Also update `ujust toggle-anticheat-support` to use a new file at
`/etc/sysctl.d/61-ptrace-scope.conf`, and to attempt to apply the
changes immediately. This removes the need for a restart when setting
ptrace_scope to 3, but not when setting ptrace_scope to 1, since
ptrace_scope=3 persists until restart.

Finally, the audit script no longer needs to check whether the sysctl
conf file has been modified, since it'll now be in `/usr` and therefore
read-only.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>