fix: user service for security update notifications (#1451)
Currently the security update notification doesn't work because the
notify-send is run as root. Fix this by invoking the script as a user
service instead, which is activated by a path unit triggered by changes
to a file that `rpm-ostreed-automatic.service` writes to. Also move the
script and associated service to the "desktop" folder since it's not
applicable to securecore.

Also make "moderate"-severity security advisories a non-critical
notification; generally these are security issues whose impact is
limited to denial-of-service or are only potentially exploitable under
unusual circumstances, so I think the less obtrusive category of
notification is more appropriate here.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>