fix: detect non-system users by presence of /var/home/$USER (#1580)
Currently systemd user services that are only supposed to run on desktop
images for real users (not system accounts) check for the existence of a
`~/.config` directory to determine whether they should run. This works
on typical desktop installations but can fail in circumstances where the
user is created with a bare-bones home directory for whatever reason.
(For example, this is the case in the integration testing environment.)

Instead, we can just check for the existence of a `/var/home/$USER`
directory. This seems like a more reliable indicator of whether a user
account is an ordinary user as opposed to some sort of system account,
and works for bare-bones home directories just as well.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>