feat: rewrite ujust harden-flatpak in Python (#1599)
* feat: rewrite ujust harden-flatpak in Python

The Python rewrite has the same basic logic but with some improvements
to user-facing messages and the app ID search logic.

The third-party `inquirer` module is used for letting the user select
from a list of options, and is installed via RPM. (I put it in
`common-packages.yml` rather than `desktop-packages.yml` because it's
likely to be useful in other ujust scripts, and is a small package in
any case.)

Also move audit-script-specific parts of the `utils` module into a new
`audit_utils` module, and fix a minor type error in `is_using_vpn`.

* fix: use [positional-arguments] tag instead of global setting

* chore: use manuallyinstalljust.sh in tests

This ensures the version of `just` used in CI is the same as the one
used in the images. Also made some minor reliability improvements to
`manuallyinstalljust.sh`: use `mktemp` to generate a temporary directory
instead of just using `/tmp`, and set `curl` options to retry on failure
and to not suppress error message output.

* chore: text wrapping, add note about host-os:ro permission