fix: generate correct provenance on non-default branches (#1637)
The generated provenance should only be for the `latest` tag on the
default branch; otherwise, the tag has a form generated from the branch
name or PR number and the OS major version number.

Also simplified the base image verification steps in the build workflows
so they don't rely on assumptions about secureblue image naming
conventions.