fix: check for bad selinux file context (#1665)
In ostree-based images, the SELinux file context alias `/var/home =
/home` needs to be reversed to `/home = /var/home`. This is supposed to
be patched in by rpm-ostree at build time, but for unclear reasons this
sometimes is failing to take place.

We don't want this bug to silently pass into builds--it messes up file
contexts for the whole home directory--so as a stopgap until we debug
this further, let's add a script to check for this error and fail the
build if it does occur.